Reorganized JUnit dialects.
