+++ tests/unit/Unittesting/JUnit.py
@@ -32,11 +32,11 @@ from datetime import timedelta, datetime
 from pathlib  import Path
 from unittest import TestCase as py_TestCase
 
-from pyEDAA.Reports.Unittesting              import TestcaseStatus, TestsuiteStatus, TestsuiteKind
-from pyEDAA.Reports.Unittesting              import TestsuiteSummary as ut_TestsuiteSummary
-from pyEDAA.Reports.Unittesting              import Testsuite as ut_Testsuite, Testcase as ut_Testcase
-from pyEDAA.Reports.Unittesting.GenericJUnit import UnittestException
-from pyEDAA.Reports.Unittesting.GenericJUnit import Testcase, Testclass, Testsuite, TestsuiteSummary, Document as JUnitDocument
+from pyEDAA.Reports.Unittesting       import TestcaseStatus, TestsuiteStatus, TestsuiteKind
+from pyEDAA.Reports.Unittesting       import TestsuiteSummary as ut_TestsuiteSummary
+from pyEDAA.Reports.Unittesting       import Testsuite as ut_Testsuite, Testcase as ut_Testcase
+from pyEDAA.Reports.Unittesting.JUnit import UnittestException
+from pyEDAA.Reports.Unittesting.JUnit import Testcase, Testclass, Testsuite, TestsuiteSummary, Document as JUnitDocument
 
 
 class Instantiation(py_TestCase):
