+++ tests/unit/Unittesting/Merge.py
@@ -33,8 +33,8 @@ from time     import perf_counter_ns
 from typing   import List
 from unittest import TestCase as ut_TestCase
 
-from pyEDAA.Reports.Unittesting              import MergedTestsuiteSummary, IterationScheme, TestcaseStatus
-from pyEDAA.Reports.Unittesting.GenericJUnit import Document, JUnitReaderMode
+from pyEDAA.Reports.Unittesting       import MergedTestsuiteSummary, IterationScheme, TestcaseStatus
+from pyEDAA.Reports.Unittesting.JUnit import Document, JUnitReaderMode
 
 
 class PyTooling(ut_TestCase):
